Clean up SearchMethod type duplication and remove dead code
WHAT:
- Move SearchMethod type from search/types to shared/stores/UiOnlyStore.ts
- Remove duplicate SearchMethod definition from search-types.ts
- Remove unused SearchEventDetail and ViewAllEventDetail interfaces

WHY:
- SearchMethod is UI state (tracks which search method was used for display)
- Shared stores shouldn't import from feature-specific types
- Dead interfaces were never used anywhere in codebase

ARCHITECTURAL FIX:
- Before: shared/stores/UiOnlyStore imports from search/types (backwards dependency)
- After: search/components/SearchResults imports SearchMethod from shared/stores (correct flow)
- Result: Clean dependency flow Features → Shared

CLEANUP:
- Single source of truth for SearchMethod type
- Removed 2 unused interfaces (dead code)
- Added missing 'minisearch_type_selection' option to complete type

✅ TESTED: All search functionality working correctly
